+++ utils/ns_generate_acks.py
@@ -3,12 +3,13 @@
 import json
 from pathlib import Path
 
-acks_path = Path(__file__).parent.parent.absolute() / "src" / "neosca" / "ns_data" / "acks.json"
+PKG_DIR = Path(__file__).parent.parent.absolute()
+acks_path = PKG_DIR / "src" / "neosca" / "ns_data" / "acks.json"
 with open(acks_path, encoding="utf-8") as f:
     ack_data = json.load(f)
 acknowledgment, projects = ack_data.values()
 
-filepath = "../ACKNOWLEDGMENTS.md"
+filepath = PKG_DIR / "ACKNOWLEDGMENTS.md"
 with open(filepath, "w", encoding="utf-8") as f:
     title = "# Acknowledgments\n\n"
     header = "||Name|Version|Authors|License|\n"